+++ hawkbit-http-security/src/main/java/org/eclipse/hawkbit/security/DosFilter.java
