+++ hawkbit-security-core/pom.xml
@@ -32,10 +32,6 @@
          <scope>provided</scope>
       </dependency>
       <dependency>
-         <groupId>org.springframework</groupId>
-         <artifactId>spring-web</artifactId>
-      </dependency>
-      <dependency>
          <groupId>com.github.ben-manes.caffeine</groupId>
          <artifactId>caffeine</artifactId>
       </dependency>
